Refactors for ensoV4 -- the package is still WIP - drops App folder & namespace - replaces fillable with guarded id - adds type in structure migrations - replaces web middleware group with api for sanctum - updates deps
