+++ routes/api.php
@@ -1,0 +1,17 @@
+<?php
+
+use Illuminate\Support\Facades\Route;
+
+Route::middleware(['api', 'auth', 'core'])
+    ->namespace('LaravelEnso\Emails\Http\Controllers\Emails')
+    ->prefix('api/emails')
+    ->as('emails.')
+    ->group(function () {
+        Route::post('send', 'Send')->name('send');
+        Route::post('', 'Store')->name('store');
+        Route::get('{email}/edit', 'Edit')->name('edit');
+        Route::post('{email}/update', 'Update')->name('update');
+        Route::delete('{email}', 'Destroy')->name('destroy');
+        Route::get('initTable', 'InitTable')->name('initTable');
+        Route::get('tableData', 'TableData')->name('tableData');
+    });
