--- src/App/Enums/Priorities.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace LaravelEnso\Emails\App\Enums;
-
-use LaravelEnso\Enums\App\Services\Enum;
-
-class Priorities extends Enum
-{
-    public const High = 1;
-    public const Normal = 3;
-    public const Low = 5;
-}
