--- src/App/Enums/PriorityLabels.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace LaravelEnso\Emails\App\Enums;
-
-use LaravelEnso\Enums\App\Services\Enum;
-
-class PriorityLabels extends Enum
-{
-    protected static array $data = [
-        Priorities::High => 'is-danger',
-        Priorities::Normal => 'is-warning',
-        Priorities::Low => 'is-success',
-    ];
-}
