--- src/App/Http/Resources/Recipient.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace LaravelEnso\Emails\App\Http\Resources;
-
-use Illuminate\Http\Resources\Json\JsonResource;
-
-class Recipient extends JsonResource
-{
-    public function toArray($request)
-    {
-        return ['id' => $this->id];
-    }
-}
