+++ src/Enums/Priorities.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace LaravelEnso\Emails\Enums;
+
+use LaravelEnso\Enums\Services\Enum;
+
+class Priorities extends Enum
+{
+    public const High = 1;
+    public const Normal = 3;
+    public const Low = 5;
+}
