+++ src/Enums/StatusLabels.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace LaravelEnso\Emails\Enums;
+
+use LaravelEnso\Enums\Services\Enum;
+
+class StatusLabels extends Enum
+{
+    protected static array $data = [
+        Statuses::Sent => 'is-success',
+        Statuses::Scheduled => 'is-warning',
+        Statuses::Draft => 'is-info',
+    ];
+}
