+++ src/Http/Controllers/Emails/InitTable.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace LaravelEnso\Emails\Http\Controllers\Emails;
+
+use Illuminate\Routing\Controller;
+use LaravelEnso\Emails\Tables\Builders\EmailTable;
+use LaravelEnso\Tables\Traits\Init;
+
+class InitTable extends Controller
+{
+    use Init;
+
+    protected $tableClass = EmailTable::class;
+}
