+++ src/Http/Controllers/Emails/Store.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace LaravelEnso\Emails\Http\Controllers\Emails;
+
+use Illuminate\Routing\Controller;
+use LaravelEnso\Emails\Http\Requests\ValidateEmailRequest;
+use LaravelEnso\Emails\Models\Email;
+
+class Store extends Controller
+{
+    public function __invoke(ValidateEmailRequest $request, Email $email)
+    {
+        $email->fill($request->mapped())->save();
+
+        $email->syncRecipients()
+            ->syncAttachments()
+            ->send();
+
+        return [
+            'message' => __('The email was successfully saved!'),
+            'redirect' => 'emails.edit',
+            'params' => ['email' => $email->id],
+        ];
+    }
+}
