+++ src/Http/Controllers/Emails/TableData.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace LaravelEnso\Emails\Http\Controllers\Emails;
+
+use Illuminate\Routing\Controller;
+use LaravelEnso\Emails\Tables\Builders\EmailTable;
+use LaravelEnso\Tables\Traits\Data;
+
+class TableData extends Controller
+{
+    use Data;
+
+    protected $tableClass = EmailTable::class;
+}
