+++ src/Http/Resources/Recipient.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace LaravelEnso\Emails\Http\Resources;
+
+use Illuminate\Http\Resources\Json\JsonResource;
+
+class Recipient extends JsonResource
+{
+    public function toArray($request)
+    {
+        return ['id' => $this->id];
+    }
+}
