--- src/config/emails.php
@@ -1,5 +1,0 @@
-<?php
-
-return [
-    //
-];
